[New] Added test cases for pages admin roles form forms
