+++ app/javascript/components/pages/admin/roles-form/forms/associated-roles.spec.js
@@ -7,8 +7,6 @@ describe("pages/admin/<RolesForm>/forms - AssociatedRolesForm", () => {
 
   it("returns the AssociatedRolesForm with fields", () => {
     const roleForms = AssociatedRolesForm(fromJS([]), i18n);
-
-    expect(roleForms).to.exist;
-    expect(roleForms.fields).to.have.lengthOf(2);
+    expect(roleForms.fields).toHaveLength(2);
   });
 });
